+++ CHANGELOG.md
@@ -21,6 +21,7 @@ cider-nrepl's info middleware for jump-to-definition.
 * New inspector inspired by SLIME's inspector
 * STDERR ouput is now font-locked with `cider-repl-err-output-face` to make it
 visually distinctive from `cider-repl-output-face` (used for STDOUT output).
+* New interactive command `cider-scratch`.
 
 ### Changes